backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
ultralytics>=8.1.0
pillow>=10.4.0
python-multipart>=0.0.17
numpy>=1.26.0
sqlmodel>=0.0.16

# PostgreSQL
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Database migrations
alembic>=1.13.0

# File storage & processing
aiofiles>=23.2.1
python-magic>=0.4.27

# Validation & config
pydantic-settings>=2.2.0
python-dotenv>=1.0.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.17

# Logging
loguru>=0.7.2

# Utils
python-dateutil>=2.8.2
redis>=5.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

greenlet>=3.2.4